+++ client/package.json
@@ -19,7 +19,8 @@
     "react-redux": "^7.1.0",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.0.1",
-    "redux": "^4.0.4"
+    "redux": "^4.0.4",
+    "redux-saga": "^1.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
