+++ client/src/services/user.js
@@ -1,7 +1,11 @@
+import axios from 'axios';
 
 export const userService = {
-    login: () => {
-
+    login: (user) => {
+        return axios.post('/users/login', user).then(res => {
+            localStorage.setItem('token', res.data.token);
+            return res.data;
+        });
     },
     logout: () => {
